Remove deprecations from autoconfigure module
